requirements.txt
streamlit
PyPDF2
langchain
langchain-google-genai
google-generativeai
faiss-cpu
pillow
python-dotenv
langchain-community